+++ taverna-language-commandline/src/main/java/org/apache/tavlang/commandline/TavernaCommandline.java
@@ -1,4 +1,4 @@
-package org.apache.taverna.commandline;
+package org.apache.tavlang.commandline;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
